refactor: 잘못 사용한 DTO import 변경
