+++ app/src/main/java/woowacourse/shopping/data/service/ProductService.kt
@@ -4,7 +4,7 @@ import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
-import woowacourse.shopping.data.dto.cartitem.ProductResponse
+import woowacourse.shopping.data.dto.product.ProductResponse
 
 interface ProductService {
     @GET("/products")
